move code to own files.
